Issue #1787: Move ViewModelFactory to architecture pkg.
